% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_metadata.R
\name{all_metadata}
\alias{all_metadata}
\title{This function downloads the metadata for all projects.}
\usage{
all_metadata(subset = "sra", verbose = TRUE)
}
\arguments{
\item{subset}{Either \code{sra} or \code{gtex}. Specifies which metadata
file to download.}

\item{verbose}{If \code{TRUE} it will print a message of where the file is
being downloaded to.}
}
\value{
A \link[S4Vectors]{DataFrame-class} object with the phenotype
metadata.
}
\description{
Download the metadata from all the projects. This can be useful for finding
samples of interests across all projects.
}
\examples{

metadata <- all_metadata()

}
\author{
Leonardo Collado-Torres
}

