% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount-package.R
\docType{package}
\name{recount-package}
\alias{recount-package}
\title{Explore and download data from the recount project.}
\description{
Explore and download data from the recount project available at
\url{https://jhubiostatistics.shinyapps.io/recount/}. Using the recount package
you can download
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
objects at the gene or exon level, the raw counts, the phenotype metadata
used, the urls to the sample coverage bigWig files or the mean coverage
bigWig file for a particular study. The
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class} objects can be
used by different packages for performing differential expression analysis.
Using \url{http://bioconductor.org/packages/derfinder} you can perform
annotation-agnostic differential expression analyses with the data from the
recount project.
}
\author{
Leonardo Collado-Torres \href{mailto:lcolladotor@gmail.com}{lcolladotor@gmail.com}
}
\keyword{package}
