% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount_abstract-data.R
\docType{data}
\name{recount_abstract}
\alias{recount_abstract}
\title{Summary information at the project level for the recount project}
\format{A data.frame with 4 columns.
\describe{
    \item{number_samples }{ the number of samples in the study,}
    \item{species }{ the species of the study,}
    \item{abstract }{ the abstract text as provided by the SRAdb Bioconductor
    package,}
    \item{project }{ the SRA project id.}
}}
\description{
A data.frame with summary information at the project level for the studies
analyzed in the recount project.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
\seealso{
\link{download_study}
}
\keyword{datasets}

