% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount_exons-data.R
\docType{data}
\name{recount_exons}
\alias{recount_exons}
\title{Exon annotation used in recount}
\format{A \link[GenomicRanges]{GRangesList-class} with one element per 
gene. The names match the gene Gencode v25 ids.}
\description{
Exon annotation extracted from Gencode v25 (GRCh38.p7) used in recount.
Data extraced on October 12th, 2017.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
\seealso{
\link{reproduce_ranges}, \link{recount_genes}
}
\keyword{datasets}
