% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount_genes-data.R
\docType{data}
\name{recount_genes}
\alias{recount_genes}
\title{Gene annotation used in recount}
\format{A \link[GenomicRanges]{GRanges-class} with one range per gene. The 
names match their UCSC ids. The \link[GenomicRanges]{GRanges-class} has two 
metadata columns.
\describe{
    \item{gene_id }{  the UCSC ids, identical to the names.}
    \item{bp_length }{ the sum of the width of the reduced exons for that 
    given gene.}
}}
\description{
Gene annotation extracted from TxDb.Hsapiens.UCSC.hg38.knownGene used in
recount. It includes the sum of the width of the reduced exons which can be
used for normalizing the counts provided in the 
\link[SummarizedExperiment]{RangedSummarizedExperiment-class} objects.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
\seealso{
\link{reproduce_ranges}, \link{recount_exons}
}
\keyword{datasets}

