% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount_url-data.R
\docType{data}
\name{recount_url}
\alias{recount_url}
\title{Files and URLs hosted by the recount project}
\format{A data.frame with 6 columns.
\describe{
\item{path }{ the original path to the file before being uploaded,}
\item{file_name }{ the file name,}
\item{project }{ the SRA project id,}
\item{version1 }{ A logical vector indicating whether the file was
part of version 1 (reduced exons)}.
\item{version 2}{ A logical vector indicating whether the file was
updated in version 2 (disjoint exons)}. Further details in the recount
website documentation tab.
\item{url }{ the public URL for the given file.}
}}
\description{
Files and URLs as provided by the recount project. This information is used
internally in \link{download_study}.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
\seealso{
\link{download_study}
}
\keyword{datasets}
