% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_ranges.R
\name{reproduce_ranges}
\alias{reproduce_ranges}
\title{Reproduce the gene or exons used in the RangedSummarizedExperiment objects}
\usage{
reproduce_ranges(level = "gene")
}
\arguments{
\item{level}{Either \code{genes} or \code{exon}. It specifies whether to
return Gene or exon level information as a 
\link[GenomicRanges]{GRanges-class} or 
\link[GenomicRanges]{GRangesList-class} object respectively. The gene level
information contains the width of the reduced exons for that given gene
which can be used to normalize the counts provided by recount.}
}
\value{
Either a \link[GenomicRanges]{GRanges-class} object like 
\link{recount_genes} or a \link[GenomicRanges]{GRangesList-class} object 
like \link{recount_exons}.
}
\description{
This function reproduces the gene or exon level information used for creating
the \link[SummarizedExperiment]{RangedSummarizedExperiment-class}
objects provided by recount. The annotation is based on
\code{TxDb.Hsapiens.UCSC.hg38.knownGene} with the gene-level 
information extracted with \code{genes()} (see 
\link[GenomicFeatures]{transcripts} with default arguments.
}
\details{
Note that the information used in recount was originally calculated with
\url{https://github.com/nellore/runs/blob/master/recount2/genes/create_bed.R}
with the versions as noted in
\url{https://github.com/nellore/runs/blob/master/recount2/genes/logs/genes-hg38-ucsc.o1192789}.
}
\examples{

## Reproduce gene level information
genes <- reproduce_ranges()

## Compare against recount_genes
length(genes)
length(recount_genes)

}
\author{
Leonardo Collado-Torres
}
\seealso{
\link{recount_genes}, \link{recount_exons},
    \url{https://github.com/nellore},
    \url{https://lcolladotor.shinyapps.io/recount/}
}

