% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse_gene_SRP009615-data.R
\docType{data}
\name{rse_gene_SRP009615}
\alias{rse_gene_SRP009615}
\title{RangedSummarizedExperiment at the gene level for study SRP009615}
\format{A \link[SummarizedExperiment]{RangedSummarizedExperiment-class} as 
created by the recount project for study with SRA id (accession number) 
SRP009615.}
\description{
\link[SummarizedExperiment]{RangedSummarizedExperiment-class} at the gene 
level for study SRP009615. Used as an example in \link{scale_counts}.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
\seealso{
\link{scale_counts}, \link{download_study}
}
\keyword{datasets}
