% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add additional curated metadata to a recount rse object}
\usage{
add_metadata(rse, source = "recount_brain_v1", is_tcga = FALSE,
  verbose = TRUE)
}
\arguments{
\item{rse}{A \link[SummarizedExperiment]{RangedSummarizedExperiment-class} 
object as downloaded with \link{download_study}.}

\item{source}{A valid source name. The only supported option at this
moment is \code{recount_brain_v1}.}

\item{is_tcga}{Set to \code{TRUE} only when \code{rse} is from TCGA.
Otherwise set to \code{FALSE} (default).}

\item{verbose}{If \code{TRUE} it will print a message of where the 
predictions file is being downloaded to.}
}
\value{
A \link[SummarizedExperiment]{RangedSummarizedExperiment-class} 
object with the sample metadata columns appended to the \code{colData()}
slot.
For \code{source = "recount_brain_v1"}, the metadata columns are
described at \url{http://lieberinstitute.github.io/recount-brain/}.
}
\description{
This function appends sample metadata information to a 
\link[SummarizedExperiment]{RangedSummarizedExperiment-class} from the
recount2 project. The sample metadata comes from curated efforts
independent from the original recount2 project. Currently the only
information comes from the recount_brain project described in more detail
at \url{http://lieberinstitute.github.io/recount-brain/}.
}
\details{
If you use the recount_brain data please cite the Razmara et al
bioRxiv pre-print available at 
(TODO update URL once it's available). See citation
details with citation('recount').
}
\examples{

## Add the sample metadata to an example rse_gene object
rse_gene <- add_metadata(rse_gene_SRP009615, 'recount_brain_v1')

## Explore the metadata
colData(rse_gene)

## For a list of studies present in recount_brain_v1 check
## http://lieberinstitute.github.io/recount-brain/. Note that it only
## includes studies from SRA, so no TCGA or GTEx (those have great
## sample metadata already available).

\dontrun{
## Example project that is present in recount_brain_v1.

## Download and load the data
download_study('ERP001304')
load(file.path('ERP001304', 'rse_gene.Rdata'))

## Add the sample metadata from recount_brain_v1
rse_gene <- add_metadata(rse_gene)

## Check the metadata
colData(rse_gene)
}

}
\references{
Razmara et al, in prep, 2018.
}
\author{
Leonardo Collado-Torres
}
