% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_study.R
\name{download_study}
\alias{download_study}
\title{Download data for a given SRA study id from the recount project}
\usage{
download_study(project, type = "rse-gene", outdir = project,
  download = TRUE, ...)
}
\arguments{
\item{project}{A character vector with one SRA study id.}

\item{type}{Specifies which files to download. The options are:
\describe{
    \item{rse-gene}{ the gene-level 
    \link[SummarizedExperiment]{RangedSummarizedExperiment-class} object in 
    a file named rse_gene.Rdata.}
    \item{rse-exon}{ the exon-level 
    \link[SummarizedExperiment]{RangedSummarizedExperiment-class} object in 
    a file named rse_exon.Rdata.}
    \item{counts-gene}{ the gene-level counts in a tsv file named
    counts_gene.tsv.gz.}
    \item{counts-exon}{ the exon-level counts in a tsv file named
    counts_exon.tsv.gz.}
    \item{phenotype}{ the phenotype data for the study in a tsv file named
    \code{project}.tsv.}
    \item{files-info}{ the files information for the given study (including
    md5sum hashes) in a tsv file named files_info.tsv.}
    \item{samples}{ one bigWig file per sample in the study.}
    \item{mean}{ one mean bigWig file for the samples in the study, 
    with each sample normalized to a 40 million 100 bp library using the
    total coverage sum (area under the coverage curve, AUC) for the given 
    sample.}
}}

\item{outdir}{The destination directory for the downloaded file(s).}

\item{download}{Whether to download the files or just get the download urls.}

\item{...}{Additional arguments passed to \link[utils]{download.file}.}
}
\value{
Returns invisibly the URL(s) for the files that were downloaded.
}
\description{
Download the gene or exon level 
\link[SummarizedExperiment]{RangedSummarizedExperiment-class} objects 
provided by the recount project. Alternatively download the counts, metadata 
or file information for a given SRA study id. You can also download the 
sample bigWig files or the mean coverage bigWig file.
}
\examples{
## Find the URL to download the RangedSummarizedExperiment for the
## Geuvadis consortium study.
url <- download_study('ERP001942', download = FALSE)

## See the actual URL
url

## Download the example data included in the package for study SRP009615
\dontrun{
url2 <- download_study('SRP009615')
url2

## Load the data
load(file.path('SRP009615', 'rse_gene.Rdata'))

## Compare the data
identical(rse_gene, rse_gene_SRP009615)
}
}
\author{
Leonardo Collado-Torres
}

