% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_info.R
\name{geo_info}
\alias{geo_info}
\title{Extract information from GEO for a given sample}
\usage{
geo_info(geoid, verbose = FALSE)
}
\arguments{
\item{geoid}{A character vector of length 1 with the GEO accession id for
a given sample.}

\item{verbose}{If \code{TRUE}, messages from \link[GEOquery]{getGEO} will
be shown. Otherwise they are suppressed.}
}
\value{
Returns a \link[S4Vectors]{DataFrame-class} with the information 
from GEO available for the given sample.
}
\description{
This function uses GEOquery to extract information for a given sample. The
GEO accession ids for the sample can be found in the study phenotype table.
}
\examples{
geo_info('GSM836270')

}
\author{
Leonardo Collado-Torres, Andrew Jaffe
}

