% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount_exons-data.R
\docType{data}
\name{recount_exons}
\alias{recount_exons}
\title{Exon annotation used in recount}
\format{A \link[GenomicRanges]{GRangesList-class} with one element per 
gene. The names match the gene UCSC ids.}
\description{
Exon annotation extracted from TxDb.Hsapiens.UCSC.hg38.knownGene used in
recount.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
\seealso{
\link{reproduce_ranges}, \link{recount_genes}
}
\keyword{datasets}

