% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recount_url-data.R
\docType{data}
\name{recount_url}
\alias{recount_url}
\title{Files and URLs hosted by the recount project}
\format{A data.frame with 4 columns.
\describe{
    \item{path }{ the original path to the file before being uploaded,}
    \item{file_name }{ the file name,}
    \item{project }{ the SRA project id,}
    \item{url }{ the public URL for the given file.}
}}
\description{
Files and URLs as provided by the recount project. This information is used
internally in \link{download_study}.
}
\references{
\url{https://jhubiostatistics.shinyapps.io/recount/}
}
\seealso{
\link{download_study}
}
\keyword{datasets}

