% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_ranges.R
\name{reproduce_ranges}
\alias{reproduce_ranges}
\title{Reproduce the gene or exons used in the RangedSummarizedExperiment objects}
\usage{
reproduce_ranges(level = "gene", db = "TxDb.Hsapiens.UCSC.hg38.knownGene")
}
\arguments{
\item{level}{Either \code{genes} or \code{exon}. It specifies whether to
return Gene or exon level information as a 
\link[GenomicRanges]{GRanges-class} or 
\link[GenomicRanges]{GRangesList-class} object respectively. The gene level
information contains the width of the reduced exons for that given gene
which can be used to normalize the counts provided by recount.}

\item{db}{Either \code{TxDb.Hsapiens.UCSC.hg38.knownGene} (default) or
\code{EnsDb.Hsapiens.v79}. The default option reproduces the annotation
used when creating recount (when using TxDb.Hsapiens.UCSC.hg38.knownGene
version 3.1.3) or updates it. EnsDb.Hsapiens.v79 can be used
for an alternative annotation as showcased in the recount vignette.}
}
\value{
Either a \link[GenomicRanges]{GRanges-class} object like 
\link{recount_genes} or a \link[GenomicRanges]{GRangesList-class} object 
like \link{recount_exons}.
}
\description{
This function reproduces the gene or exon level information used for creating
the \link[SummarizedExperiment]{RangedSummarizedExperiment-class}
objects provided by recount. The annotation is based on
\code{TxDb.Hsapiens.UCSC.hg38.knownGene} with the gene-level 
information extracted with \code{genes()} (see 
\link[GenomicFeatures]{transcripts} with default arguments. It can also be
used to generate similar results with newer annotation or with alternative
annotations for the human hg38 assembly.
}
\details{
Note that the information used in recount was originally calculated with
\url{https://github.com/leekgroup/recount-website/blob/master/genes/create_bed.R}
with the versions as noted in
\url{https://github.com/leekgroup/recount-website/blob/master/genes/logs/genes-hg38-ucsc.o1192789}.
The actual package tarball for TxDb.Hsapiens.UCSC.hg38.knownGene we used is 
available at \url{https://jhubiostatistics.shinyapps.io/recount/TxDb.Hsapiens.UCSC.hg38.knownGene_3.1.3.tar.gz}.
}
\examples{

## Reproduce gene level information
genes <- reproduce_ranges()

\dontrun{
## Compare against recount_genes
length(genes)
length(recount_genes)
}

}
\author{
Leonardo Collado-Torres
}
\seealso{
\link{recount_genes}, \link{recount_exons},
    \url{https://github.com/nellore},
    \url{https://jhubiostatistics.shinyapps.io/recount/}
}

