% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snaptron_query.R
\name{snaptron_query}
\alias{snaptron_query}
\title{Query Snaptron to get data from exon-exon junctions present in Intropolis}
\usage{
snaptron_query(junctions, version = 1, verbose = TRUE)
}
\arguments{
\item{junctions}{A \link[GenomicRanges]{GRanges-class} object with the
exon-exon junctions of interest. The chromosome names should be in UCSC
format, such as 'chr1'. The strand information is ignored in the query.}

\item{version}{Either \code{1} or \code{2}. Version 1 of Intropolis has the
exon-exon junctions from about 20 thousand RNA-seq samples in hg19 
coordinates. Version 2 has the data from about 50 thousand RNA-seq samples #' aligned to hg38.}

\item{verbose}{If \code{TRUE} status updates will be printed.}
}
\value{
A \link[GenomicRanges]{GRanges-class} object with the results from
the Snaptron query. For information on the different columns please see
\url{http://stingray.cs.jhu.edu:8090/snaptron/docs/}.
}
\description{
This function uses the Snaptron API to query specific exon-exon junctions
that are available via Intropolis as described in the vignette.
}
\examples{

library('GenomicRanges')
junctions <- GRanges(seqnames = 'chr2', IRanges(
    start = c(28971710:28971712, 29555081:29555083, 29754982:29754984),
    end = c(29462417:29462419, 29923338:29923340, 29917714:29917716)))

snaptron_query(junctions)

}
\author{
Leonardo Collado-Torres
}
\references{
Please cite \url{http://stingray.cs.jhu.edu:8090/snaptron/docs/}
if you use this function as Snaptron is a separate project from recount.
Thank you!
}

